hashers
Module fnv

The Fowler–Noll–Vo hash function.

http://www.isthe.com/chongo/tech/comp/fnv/

The basis of the FNV hash algorithm was taken from an idea sent as reviewer comments to the IEEE POSIX P1003.2 committee by Glenn Fowler and Phong Vo back in 1991. In a subsequent ballot round: Landon Curt Noll improved on their algorithm. Some people tried this hash and found that it worked rather well. In an EMail message to Landon, they named it the ``Fowler/Noll/Vo’’ or FNV hash.

FNV hashes are designed to be fast while maintaining a low collision rate. The FNV speed allows one to quickly hash lots of data while maintaining a reasonable collision rate. The high dispersion of the FNV hashes makes them well suited for hashing nearly identical strings such as URLs, hostnames, filenames, text, IP addresses, etc.

The IETF has an informational draft on The FNV Non-Cryptographic Hash Algorithm

This module provides both 32- and 64-bit versions of FNV-1a.

Structs

FNV1aHasher32

FNV1aHasher64

Functions

fnv1a32

Provide access to FNV1aHasher32 in a single call.

fnv1a64

Provide access to FNV1aHasher64 in a single call.